This changeset moves (graduates) itmf/qtff namespace into impl so that it is formally available to all libmp4v2 implementation.
The other major component of this changeset is that it undoes ugly sinit()
and sshutdown() requirement of libutil. This solution was put in place as a
result of Cygwin static initialization issues. Unknown at the time, the issue
was related to undefined behavior of extern "C" symbol placement inside
a C++ namespace and was identified during r159 Cygwin fix. As it turns out,
the same undefined behavior caused static-initialization to be skipped on
the Cygwin platform. Thus, we're now able to get rid of this ugly dynamic
sinit/sshutdown pattern; classes and code in both itmf/qtff were more or
less the central location for these changes.

And finally, a class template implementation has been put in place for enum
to/from string conversions. This means we use a single code base for all
the conversions, and the behavior is consistent. It is now exposed to
any C++ implementation code, which opens the door, potentially, to exposing
our now excellent iTMF BasicType, GenreType and StikType with C-wrappers
to the public API.

OSX and Cygwin platforms were used to test/verify static initialization
sanity and basic operation of utilities.

- moved namespace util::itmf -> impl::itmf.
- moved namespace util::qtff -> impl::qtff.
- moved util::UtilException -> impl::MP4Exception.
- added src/enum.h/.tcc for Enum class template.
- BasicType, GenreType, StikType all converted to use template.
- all utilities updated as appropriate:
    1. removal of ssinit/sshutdown requirement.
    2. use of Enum classes BasicType, GenreType and StikType.
- added src/text.h/.cpp to support case-insensitive C++ string sorting,
  currently only used by Enum template but is useful for any STL map or set
  with keys of string-type.
- minor enhancement to mp4art to attempt to determine image type when implicit.

TODO: VStudio build needs updating.